changes to make compilation with -Wall possible
git-svn-id: https://svn.nfit.au.dk/svn/beta/trunk@1344 678b6752-a8a0-49ff-a511-7c2a3241fcd4

+++ betarun/GC/stack.c
@@ -938,8 +938,8 @@ void PrintAR(struct RegWin *ar, struct RegWin *theEnd)
 
   fprintf(output, 
 	  "----- AR: 0x%x, theEnd: 0x%x, PC: 0x%x (%s+0x%x)\n",
-	  ar, 
-	  theEnd,
+	  (int)ar, 
+	  (int)theEnd,
 	  PC,
 	  getLabel(PC),
 	  labelOffset);
